Mountain events review
Again, a lot of moving events over to general. Some of them were dupes of events in Forest, so I removed them from both biome-specific jsons and put them in general. As well, I found a few deaths that existed in general but this time specifically targeted leads, so I deleted them and edited the general deaths to include leaders.