Plains events review
As it turns out, none of the existing plains/mountains events are biome specific. I remember, Scribble, you asked for the burrow events to be labelled plains -- I added beach and forest, thinking that it made sense for the other biomes too, but if you want me to relegate them to the plains.json so that it has something unique, I can do that.